+++ pyblish_maya/service.py
@@ -1,6 +1,5 @@
 # Dependencies
 import pyblish_endpoint.service
-from .version import version
 
 from maya import utils
 
@@ -12,11 +11,6 @@ class MayaService(pyblish_endpoint.service.EndpointService):
         orig = super(MayaService, self).init
         return wrapper(orig, *args, **kwargs)
 
-    def next(self, *args, **kwargs):
-        orig = super(MayaService, self).next
+    def process(self, *args, **kwargs):
+        orig = super(MayaService, self).process
         return wrapper(orig, *args, **kwargs)
-
-    def versions(self):
-        versions = super(MayaService, self).versions()
-        versions["pyblish-nuke"] = version
-        return versions
